+++ launcher.sh
@@ -6,11 +6,11 @@
 
 launch()
 {
-    # echo "0.1. Decompose each document in dataset into paragraphs."
-    # python -m modules.data_reading.data_reading --num_proc 5
+    # echo "1. Decompose each document in dataset into paragraphs."
+    python -m modules.data_reading.data_reading --num_proc 5
 
-    # echo "0.2. Find golden passages for each question and trigger training of ParasSelection"
-    python -m modules.paras_selection.paras_selection --num_proc 5
+    # echo "2. Start training"
+    python -m modules.narrativepieline.NarrativePipeline  --num_proc 5
 }